cap on pool distribution added, removing incentive to execute unprofitable jobs and drain pool

+++ test/mocks/MockCoordinator.sol
@@ -60,7 +60,7 @@ contract MockCoordinator is Coordinator {
         epoch = _epoch;
     }
 
-    function setNumberOfActiveExecutors(uint40 _numberOfActiveExecutors) public {
+    function setNumberOfActiveExecutors(uint32 _numberOfActiveExecutors) public {
         numberOfActiveExecutors = _numberOfActiveExecutors;
     }
